requirements.txt
Scrapy
grailed_api
polars
google-generativeai
python-dotenv
jupyterlab
hopsworks[python]
ipywidgets